Moved expected_cardinal to meta/
